Login/logout changes
Added datastore query to login
Added session termination on logout

loginCheck (in levr_utils) will now return an object called headerData
we'll use this to switch the way that the header is drawn, and to add
the contact_owner name to the header

i'm going to go smoke your last flattened cigarette you left on my
table. cheers!