Added login-dependant HTML. Going to bed.
loginCheck has been extended a bit to send back an object called
headerData. It now has two modes:
If strict=True, redirects to login (no headerData)
If strict=False, returns headerData without redirecting. This is useful
on pages that can be accessed by both logged-in and non-logged-in
merchants, like the landing page (/merchants) and the login page (will
add dynamic headers to this later)